Initial Release of CurveFit.jl
This functionality was separated from Optim.jl, with an attempt to
preserve history of the files using a the `git filter-branch` capability.

This represents the minimal changes to the original code to make this
a working stand-alone package, including the relevant tests and
documentation (README). More clean-up and separation still required.