Revert some of the changes of Optim.jl PR #157.
Optim.jl PR #157 included a number of changes for potential performance
improvement to our implementation of Levenberg-Marquardt. However, these changes
came at a significant cost to code readability. Consequently, I decided to
measure the efficacy of the changes. With the exception of the revised calculation
of delta_x, none of the changes had a measurable impact. The change to the
delta_x computation improves performance by 2-5%.