Geomerged (#117)
* Half baked half working geodesic acceleration

* Adding test/lm.jl back to tests

* Changed focus point from h! to avv, allowing more flexibility, and in the end better performance
I'm not too sure what to do with vHv! now. It's probably practical to abstract the problem away,
but it is really terrible in terms of performance. I think it should be eliminated and replaced
with simply a good and careful entry in the readme

* Added comments to the code

* Had forgotten to update test/geodesic.jl

* changed geodesic structure, now avv! is an optional argument.
It largely simplifies the code.
Added weighted case for tests too

* Slight refactoring, more in the spirit of Mark's code
It matches exactly his results from his test case

* clean up a bit, and try to resolve merging conflict

* yet another try at solving conflict

* small clean up

* Update the README to take into account the geodesic accel

* Better readme, geodesic section

* Added an tiny API allowing to create avv! automatically

* Removed BenchmarkTools in geodesic.jl

* Removed some "Function" type annotations, moved the variable mutated to first place in a couple of functions

* Clean.

* fix keyword separator in lmit with four args