Specify that the model parameters must be Floats
I think this should be specified. I wanted to use the curve_fit function and I followed the tutorial but in my use case I was getting an Inexact error Int64(Float64). It took me about an hour to figure out this simple mistake. Model parameters must be floats. Is it best to specify it in order to make this make easier to use for new users and avoid frustration.