Fix argument order in example
The example for the function Avv! which calculates the directional derivative does not have the same argument order as in the source code and therefore does not work.

Also, I believe that for many applications it is worthwhile to use automatic differentiation for the computation of the directional derivative even if this is less efficient. I took the liberty of also adding an example for how such an implementation might look with Zygote.