Update curve_fit.jl
An approach to solve the problem here:
https://github.com/JuliaNLSolvers/LsqFit.jl/pull/154

It would be easier and cleaner to remove the modification of check_data_health and modify
p0::AbstractVector{<:Real}
to
p0::AbstractVector{<:AbstractFloat}

But that could be restricting for users. You have the freedom to decide ;)

BTW: levenberg_marquardt takes a Vector, which is why I have chainged p0::AbstractArray to AbstractVector. {<:Real} is to make sure that we have at least a real number (which could be an integer).