Use StatsAPI instead of StatsBase, define `confint` and `vcov`
StatsAPI provides a common namespace for packages providing statistical
functionality. It supersedes StatsBase's use for this purpose.

As part of the migration to using StatsAPI, which is itself
non-functional, I've made two functional changes in the form of
deprecations of custom functions in favor of their existing StatsAPI
equivalents. Namely, `estimate_covar` has been deprecated in favor of
`vcov` and `confidence_interval` has been deprecated in favor of
`confint`. While the former is a simple renaming, the latter also
involves specifying the confidence level rather than the Type I error
rate (1 - alpha instead of alpha).