Support StaticArrays
This package is almost ready to be used straight up with StaticArrays.
Most variables are already written in a general way, there are just two
blockers:
The preallocated buffers `JJ` and `n_buffer`, aswell as the LMResults
type.

This commit fixes that, and almost allows usage of StaticArrays.
However, there's still
https://github.com/JuliaDiff/DiffResults.jl/issues/25, but that can be
currently overcome by two typedefs.