Fix documentation deployment (#247)
There was no compatibility set for the Documenter dependency in
`docs/Project.toml`, so a breaking change made in Documenter 1.0 broke
documentation deployment for this package. In particular, the `strict`
keyword argument in `makedocs` was replaced with `warnonly`. To fix it,
we can declare explicit compatibility with Documenter 1.0 and switch to
using `warnonly=true` rather than `strict=false`.

As part of this, I've also added standard entries to `.gitignore`, two
of which are relevant to documentation: the automatically generated
`docs/build/` directory and `Manifest.toml`.